feat: add metadata and credential schemas for GitHub, GitLab, Jira, and Redmine providers
