+++ src/services/issue-provider/registry.ts
@@ -1,8 +1,8 @@
-import { githubMetadata } from "./github";
-import { gitlabMetadata } from "./gitlab";
-import { jiraMetadata } from "./jira";
+import { githubMetadata } from "./github.metadata";
+import { gitlabMetadata } from "./gitlab.metadata";
+import { jiraMetadata } from "./jira.metadata";
 import type { ProviderMetadata } from "./metadata";
-import { redmineMetadata } from "./redmine";
+import { redmineMetadata } from "./redmine.metadata";
 
 /** Module-level singleton registry mapping provider type string to metadata. */
 const registry = new Map<string, ProviderMetadata>();
